DA preparation : obsope_module added & obs_module fixed
FossilOrigin-Name: b5044411ac86e60ac0018ca25d62a2926de2584e3b6c8c422dd653e0aad8524b

+++ MSM-Tactical/dpac/src/common_modules/norm_module.f90
@@ -21,7 +21,7 @@ subroutine calc_te(u,v,t,q,ps,epsq,clat,si,nlon,nlat,te)
   real(kind=dp), intent(in) :: u(:,:,:),v(:,:,:),t(:,:,:),q(:,:,:)
   real(kind=dp), intent(in) :: ps(:,:)
   real(kind=dp), intent(in) :: epsq ! weight for moist term
-  real(kind=sp), intent(in) :: clat(:),si(:)
+  real(kind=dp), intent(in) :: clat(:),si(:)
   real(kind=dp), intent(out):: te(4)
   ! for energy calculation
   integer, parameter :: kmax=21
